fix the install and pybind bug
