+++ imgui.ini
@@ -1,0 +1,16 @@
+[Window][Debug##Default]
+Pos=60,60
+Size=400,400
+
+[Window][ImPlot Demo]
+Pos=50,50
+Size=600,750
+
+[Window][Hello, world!]
+Pos=1241,53
+Size=339,180
+
+[Window][Dear ImGui Demo]
+Pos=672,50
+Size=550,680
+
