+++ src/app/app.cpp
@@ -1,6 +1,10 @@
-#include app.h"
+#include "app.h"
 #include <iostream>
 
+void App::setMessage() {
+    message = "wellcome to the app";
+}
 void App::sendMessage() {
     std::cout << message << std::endl;
 }
+
