+++ src/main_bakup.cpp
@@ -1,0 +1,8 @@
+#include "app.h"
+
+int main() {
+  App app;
+  app.setMessage();
+  app.sendMessage();
+  return 0;
+}
