+++ vcpkg.json
@@ -6,7 +6,9 @@
       "glad" ,
       "opengl-registry",
       "implot",
+      "python3",
       "pybind11",
+
     {
       "name": "imgui",
       "features": ["glfw-binding", "opengl3-binding"]
